+++ js/getRandomNumber.js
@@ -1,0 +1,4 @@
+
+export function getRandomNumber (min,max) {
+  return Math.round(Math.random() * (max - min) + min);
+}
